AICore: Human players are now considered to be like the best bot type present.
The AI has always evaluated human opponents to be like deadly bots.
Unfortunately this means that all bots will target the human players if no
deadly bots are present.

To change this the constructor now goes looking for the best ai level present
in the game and from there on considers humans to be of equal type.